Backend/requirements.txt
celery==5.3.6
channels==4.0.0
daphne==4.0.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
redis==5.0.1
psycopg2==2.9.9
google-generativeai
textblob
django-cors-headers
